Scheduling
The scheduling implementation required a major refactor of the execution
code.  Quartz is now used to handle all job execution and the
architecture has been simplified.

Additional changes:
Refactored the RestController to make use of generics.  There are now
separate RestControllers for zones and programs
Factored a ScheduleManager out of the ActivityManager.  This better
separates concerns.
Implemented several rest endpoints for scheduling Programs and Zones.
Updated the urls in the front-end to be more dynamic.